START: major: FunctionRunner and WorkMailbox refactor
